<mfai> move get_resnet_encoder() to resnet.py
