+++ mfai/torch/models/__init__.py
@@ -3,11 +3,7 @@ import pkgutil
 from pathlib import Path
 from typing import Optional, Tuple
 
-import torch.utils.model_zoo as model_zoo
 from torch import nn
-from torchvision.models.resnet import BasicBlock, Bottleneck
-
-from mfai.torch.models.resnet import ResNetEncoder
 
 from .base import AutoPaddingModel, ModelABC
 
@@ -66,81 +62,3 @@ def load_from_settings_file(
     return model_kls(
         in_channels, out_channels, input_shape=input_shape, settings=model_settings
     )
-
-
-##########################################################################################################
-######################################         Encoders           ########################################
-##########################################################################################################
-
-
-ENCODERS_MAP = {
-    "resnet18": {
-        "encoder": ResNetEncoder,
-        "pretrained_url": "https://dl.fbaipublicfiles.com/semiweaksupervision/model_files/semi_supervised_resnet18-d92f0530.pth",  # noqa
-        "params": {
-            "out_channels": (3, 64, 64, 128, 256, 512),
-            "block": BasicBlock,
-            "layers": [2, 2, 2, 2],
-        },
-    },
-    "resnet34": {
-        "encoder": ResNetEncoder,
-        "pretrained_url": "https://download.pytorch.org/models/resnet34-b627a593.pth",
-        "params": {
-            "out_channels": (3, 64, 64, 128, 256, 512),
-            "block": BasicBlock,
-            "layers": [3, 4, 6, 3],
-        },
-    },
-    "resnet50": {
-        "encoder": ResNetEncoder,
-        "pretrained_url": "https://dl.fbaipublicfiles.com/semiweaksupervision/model_files/semi_supervised_resnet50-08389792.pth",  # noqa
-        "params": {
-            "out_channels": (3, 64, 256, 512, 1024, 2048),
-            "block": Bottleneck,
-            "layers": [3, 4, 6, 3],
-        },
-    },
-}
-
-
-def get_vision_encoder(
-    name: str,
-    in_channels: int = 3,
-    depth: int = 5,
-    weights: bool = True,
-    output_stride: int = 32,
-):
-    """
-    Return an encoder with pretrained weights or not.
-    """
-    try:
-        Encoder = ENCODERS_MAP[name]["encoder"]
-    except KeyError:
-        raise KeyError(
-            "Wrong encoder name `{}`, supported ENCODERS_MAP: {}".format(
-                name, list(ENCODERS_MAP.keys())
-            )
-        )
-
-    params = ENCODERS_MAP[name]["params"]
-    params.update(depth=depth)
-    encoder = Encoder(**params)
-
-    if weights:
-        url = ENCODERS_MAP[name]["pretrained_url"]
-        pretrained = True
-        if url is None:
-            pretrained = False
-            raise KeyError(
-                f"No url is available for the pretrained encoder choosen ({name})."
-            )
-        encoder.load_state_dict(model_zoo.load_url(url))
-    else:
-        pretrained = False
-
-    encoder.set_in_channels(in_channels, pretrained=pretrained)
-    if output_stride != 32:
-        encoder.make_dilated(output_stride)
-
-    return encoder
