+++ mfai/torch/models/half_unet.py
@@ -1,16 +1,15 @@
 from collections import OrderedDict
 from dataclasses import dataclass
 from functools import reduce
-from typing import Tuple, Union
 from math import ceil
+from typing import Tuple, Union
 
 import torch
 from dataclasses_json import dataclass_json
 from torch import nn
 
-from mfai.torch.models.base import ModelABC, AutoPaddingModel, ModelType
+from mfai.torch.models.base import AutoPaddingModel, ModelABC, ModelType
 from mfai.torch.models.utils import AbsolutePosEmdebding
-
 
 
 @dataclass_json
@@ -193,7 +192,7 @@ class HalfUNet(ModelABC, AutoPaddingModel, nn.Module):
 
     def forward(self, x):
         x, old_shape = self._maybe_padding(data_tensor=x)
-        
+
         enc1 = self.encoder1(x)
         enc2 = self.encoder2(self.pool1(enc1))
         enc3 = self.encoder3(self.pool2(enc2))
@@ -207,7 +206,7 @@ class HalfUNet(ModelABC, AutoPaddingModel, nn.Module):
         )
         dec = self.decoder(summed)
         out = self.activation(self.outconv(dec))
-        
+
         return self._maybe_unpadding(out, old_shape=old_shape)
 
     @staticmethod
@@ -293,19 +292,18 @@ class HalfUNet(ModelABC, AutoPaddingModel, nn.Module):
                 *layers,
             )
         return layers
-    
-    def validate_input_shape(self, input_shape: torch.Size) -> Tuple[bool | torch.Size]:
 
-        number_pool_layers = sum(1 for layer in self.modules() if isinstance(layer, nn.MaxPool2d))
-    
-        # The UNet has M max pooling layers of size 2x2 with stride 2, each of which halves the 
-        # dimensions. For the residual connections to match shape, the input dimensions should 
+    def validate_input_shape(self, input_shape: torch.Size) -> Tuple[bool | torch.Size]:
+        number_pool_layers = sum(
+            1 for layer in self.modules() if isinstance(layer, nn.MaxPool2d)
+        )
+
+        # The UNet has M max pooling layers of size 2x2 with stride 2, each of which halves the
+        # dimensions. For the residual connections to match shape, the input dimensions should
         # be divisible by 2^N
         d = 2**number_pool_layers
-        
-        
-        new_shape = [d * ceil(input_shape[i]/d)  for i in range(len(input_shape))]
-        new_shape = torch.Size(new_shape)
 
+        new_shape = [d * ceil(input_shape[i] / d) for i in range(len(input_shape))]
+        new_shape = torch.Size(new_shape)
 
         return new_shape == input_shape, new_shape
